# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PagoPa Cruscotto Backend API",
    "description": "PagoPa Cruscotto Backend API documentation",
    "termsOfService": "",
    "license": {
      "name": "unlicensed",
      "url": ""
    },
    "version": "0.6.3"
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/modules": {
      "get": {
        "tags": [
          "anag-module-resource"
        ],
        "operationId": "getAllModules",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleDTO"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "anag-module-resource"
        ],
        "operationId": "updateModule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleRequestBean"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "anag-module-resource"
        ],
        "operationId": "createAModule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleRequestBean"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/instances": {
      "get": {
        "tags": [
          "instance-resource"
        ],
        "operationId": "getAllInstances",
        "parameters": [
          {
            "name": "partnerId",
            "in": "query",
            "required": false,
            "schema": {
              "pattern": "^[0-9]{0,25}$",
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "BOZZA",
                "PIANIFICATA",
                "IN_ESECUZIONE",
                "ESEGUITA"
              ]
            }
          },
          {
            "name": "predictedAnalysisStartDate",
            "in": "query",
            "required": false,
            "schema": {
              "pattern": "(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[012])/[0-9]{4}",
              "type": "string"
            }
          },
          {
            "name": "predictedAnalysisEndDate",
            "in": "query",
            "required": false,
            "schema": {
              "pattern": "(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[012])/[0-9]{4}",
              "type": "string"
            }
          },
          {
            "name": "analysisStartDate",
            "in": "query",
            "required": false,
            "schema": {
              "pattern": "(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[012])/[0-9]{4}",
              "type": "string"
            }
          },
          {
            "name": "analysisEndDate",
            "in": "query",
            "required": false,
            "schema": {
              "pattern": "(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[012])/[0-9]{4}",
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstanceDTO"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "instance-resource"
        ],
        "operationId": "updateInstance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceRequestBean"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "instance-resource"
        ],
        "operationId": "createAInstance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceRequestBean"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/modules/{id}": {
      "get": {
        "tags": [
          "anag-module-resource"
        ],
        "operationId": "getModule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "anag-module-resource"
        ],
        "operationId": "deleteModule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/kpi-results/b9/module/{moduleId}": {
      "get": {
        "tags": [
          "kpi-result-resource"
        ],
        "operationId": "getKpiB9Results",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiB9ResultDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-results/b2/module/{moduleId}": {
      "get": {
        "tags": [
          "kpi-result-resource"
        ],
        "operationId": "getKpiB2Results",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiB2ResultDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-results/a2/module/{moduleId}": {
      "get": {
        "tags": [
          "kpi-result-resource"
        ],
        "operationId": "getKpiA2Results",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiA2ResultDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-results/a1/module/{moduleId}": {
      "get": {
        "tags": [
          "kpi-result-resource"
        ],
        "operationId": "getKpiA1Results",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiA1ResultDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-detail-results/b9/module/{resultId}": {
      "get": {
        "tags": [
          "kpi-detail-result-resource"
        ],
        "operationId": "getKpiB9DetailResults",
        "parameters": [
          {
            "name": "resultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiB9DetailResultDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-detail-results/b2/module/{resultId}": {
      "get": {
        "tags": [
          "kpi-detail-result-resource"
        ],
        "operationId": "getKpiB2DetailResults",
        "parameters": [
          {
            "name": "resultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiB2DetailResultDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-detail-results/a2/module/{resultId}": {
      "get": {
        "tags": [
          "kpi-detail-result-resource"
        ],
        "operationId": "getKpiA2DetailResults",
        "parameters": [
          {
            "name": "resultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiA2DetailResultDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-detail-results/a1/module/{resultId}": {
      "get": {
        "tags": [
          "kpi-detail-result-resource"
        ],
        "operationId": "getKpiA1DetailResults",
        "parameters": [
          {
            "name": "resultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiA1DetailResultDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kpi-analytic-data/b9/module/{detailResultId}": {
      "get": {
        "tags": [
          "kpi-analytic-data-resource"
        ],
        "operationId": "getKpiB9AnalyticDataResults",
        "parameters": [
          {
            "name": "detailResultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiB9AnalyticDataDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-analytic-data/b2/module/{detailResultId}": {
      "get": {
        "tags": [
          "kpi-analytic-data-resource"
        ],
        "operationId": "getKpiB2AnalyticDataResults",
        "parameters": [
          {
            "name": "detailResultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiB2AnalyticDataDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-analytic-data/a2/module/{detailResultId}": {
      "get": {
        "tags": [
          "kpi-analytic-data-resource"
        ],
        "operationId": "getKpiA2AnalyticDataResults",
        "parameters": [
          {
            "name": "detailResultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiA2AnalyticDataDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/kpi-analytic-data/a1/module/{detailResultId}": {
      "get": {
        "tags": [
          "kpi-analytic-data-resource"
        ],
        "operationId": "getKpiA1AnalyticDataResults",
        "parameters": [
          {
            "name": "detailResultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KpiA1AnalyticDataDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/instances/{id}": {
      "get": {
        "tags": [
          "instance-resource"
        ],
        "operationId": "getInstance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "instance-resource"
        ],
        "operationId": "deleteInstance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/instance-modules/{id}": {
      "get": {
        "tags": [
          "instance-module-resource"
        ],
        "operationId": "getInstanceModuleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceModuleDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModuleRequestBean": {
        "required": [
          "allowManualOutcome",
          "analysisType",
          "code",
          "configAverageTimeLimit",
          "configEligibilityThreshold",
          "configEvaluationType",
          "configExcludePlannedShutdown",
          "configExcludeUnplannedShutdown",
          "configTolerance",
          "description",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "analysisType": {
            "type": "string",
            "enum": [
              "AUTOMATICA",
              "MANUALE"
            ]
          },
          "allowManualOutcome": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "ATTIVO",
              "NON_ATTIVO"
            ]
          },
          "configAverageTimeLimit": {
            "type": "boolean"
          },
          "configEligibilityThreshold": {
            "type": "boolean"
          },
          "configEvaluationType": {
            "type": "boolean"
          },
          "configExcludePlannedShutdown": {
            "type": "boolean"
          },
          "configExcludeUnplannedShutdown": {
            "type": "boolean"
          },
          "configTolerance": {
            "type": "boolean"
          }
        }
      },
      "ModuleDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "analysisType": {
            "type": "string",
            "enum": [
              "AUTOMATICA",
              "MANUALE"
            ]
          },
          "allowManualOutcome": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "ATTIVO",
              "NON_ATTIVO"
            ]
          },
          "configExcludePlannedShutdown": {
            "type": "boolean"
          },
          "configExcludeUnplannedShutdown": {
            "type": "boolean"
          },
          "configEligibilityThreshold": {
            "type": "boolean"
          },
          "configTolerance": {
            "type": "boolean"
          },
          "configAverageTimeLimit": {
            "type": "boolean"
          },
          "configEvaluationType": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "deletedDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "InstanceRequestBean": {
        "required": [
          "analysisPeriodEndDate",
          "analysisPeriodStartDate",
          "partnerId",
          "predictedDateAnalysis"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "partnerId": {
            "type": "string"
          },
          "predictedDateAnalysis": {
            "pattern": "(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[012])/[0-9]{4}",
            "type": "string"
          },
          "analysisPeriodStartDate": {
            "pattern": "(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[012])/[0-9]{4}",
            "type": "string"
          },
          "analysisPeriodEndDate": {
            "pattern": "(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[012])/[0-9]{4}",
            "type": "string"
          }
        }
      },
      "InstanceDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "instanceIdentification": {
            "type": "string"
          },
          "partnerId": {
            "type": "integer",
            "format": "int64"
          },
          "partnerFiscalCode": {
            "type": "string"
          },
          "partnerName": {
            "type": "string"
          },
          "predictedDateAnalysis": {
            "type": "string",
            "format": "date"
          },
          "applicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "assignedUserId": {
            "type": "integer",
            "format": "int64"
          },
          "assignedFirstName": {
            "type": "string"
          },
          "assignedLastName": {
            "type": "string"
          },
          "analysisPeriodStartDate": {
            "type": "string",
            "format": "date"
          },
          "analysisPeriodEndDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "BOZZA",
              "PIANIFICATA",
              "IN_ESECUZIONE",
              "ESEGUITA"
            ]
          },
          "lastAnalysisDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastAnalysisOutcome": {
            "type": "string",
            "enum": [
              "STANDBY",
              "OK",
              "KO"
            ]
          },
          "instanceModules": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstanceModuleDTO"
            }
          }
        }
      },
...
      "KpiB9ResultDTO": {
        "required": [
          "analysisDate",
          "eligibilityThreshold",
          "evaluationType",
          "excludePlannedShutdown",
          "excludeUnplannedShutdown",
          "instanceId",
          "instanceModuleId",
          "outcome",
          "tolerance"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "instanceId": {
            "type": "integer",
            "format": "int64"
          },
          "instanceModuleId": {
            "type": "integer",
            "format": "int64"
          },
          "analysisDate": {
            "type": "string",
            "format": "date"
          },
          "excludePlannedShutdown": {
            "type": "boolean"
          },
          "excludeUnplannedShutdown": {
            "type": "boolean"
          },
          "eligibilityThreshold": {
            "type": "number",
            "format": "double"
          },
          "tolerance": {
            "type": "number",
            "format": "double"
          },
          "evaluationType": {
            "type": "string",
            "enum": [
              "MESE",
              "TOTALE"
            ]
          },
          "outcome": {
            "type": "string",
            "enum": [
              "OK",
              "KO",
              "RUNNING",
              "STANDBY"
            ]
          }
        }
      },
      "KpiB2ResultDTO": {
        "required": [
          "analysisDate",
          "averageTimeLimit",
          "eligibilityThreshold",
          "evaluationType",
          "excludePlannedShutdown",
          "excludeUnplannedShutdown",
          "instanceId",
          "instanceModuleId",
          "outcome",
          "tolerance"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "instanceId": {
            "type": "integer",
            "format": "int64"
          },
          "instanceModuleId": {
            "type": "integer",
            "format": "int64"
          },
          "analysisDate": {
            "type": "string",
            "format": "date"
          },
          "excludePlannedShutdown": {
            "type": "boolean"
          },
          "excludeUnplannedShutdown": {
            "type": "boolean"
          },
          "eligibilityThreshold": {
            "type": "number",
            "format": "double"
          },
          "tolerance": {
            "type": "number",
            "format": "double"
          },
          "averageTimeLimit": {
            "type": "number",
            "format": "double"
          },
          "evaluationType": {
            "type": "string",
            "enum": [
              "MESE",
              "TOTALE"
            ]
          },
          "outcome": {
            "type": "string",
            "enum": [
              "OK",
              "KO",
              "RUNNING",
              "STANDBY"
            ]
          }
        }
      },
      "KpiA2ResultDTO": {
        "required": [
          "analysisDate",
          "instanceId",
          "instanceModuleId",
          "outcome",
          "tolerance"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "instanceId": {
            "type": "integer",
            "format": "int64"
          },
          "instanceModuleId": {
            "type": "integer",
            "format": "int64"
          },
          "analysisDate": {
            "type": "string",
            "format": "date"
          },
          "tolerance": {
            "type": "number",
            "format": "double"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "OK",
              "KO",
              "RUNNING",
              "STANDBY"
            ]
          }
        }
      },
      "KpiA1ResultDTO": {
        "required": [
          "analysisDate",
          "eligibilityThreshold",
          "evaluationType",
          "excludePlannedShutdown",
          "excludeUnplannedShutdown",
          "instanceId",
          "instanceModuleId",
          "outcome",
          "tolerance"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "instanceId": {
            "type": "integer",
            "format": "int64"
          },
          "instanceModuleId": {
            "type": "integer",
            "format": "int64"
          },
          "analysisDate": {
            "type": "string",
            "format": "date"
          },
          "excludePlannedShutdown": {
            "type": "boolean"
          },
          "excludeUnplannedShutdown": {
            "type": "boolean"
          },
          "eligibilityThreshold": {
            "type": "number",
            "format": "double"
          },
          "tolerance": {
            "type": "number",
            "format": "double"
          },
          "evaluationType": {
            "type": "string",
            "enum": [
              "MESE",
              "TOTALE"
            ]
          },
          "outcome": {
            "type": "string",
            "enum": [
              "OK",
              "KO",
              "RUNNING",
              "STANDBY"
            ]
          }
        }
      },
...
      "KpiB9AnalyticDataDTO": {
        "required": [
          "analysisDate",
          "evaluationDate",
          "instanceId",
          "instanceModuleId",
          "resKoReal",
          "resKoValid",
          "resOk",
          "stationId",
          "totRes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "instanceId": {
            "type": "integer",
            "format": "int64"
          },
          "instanceModuleId": {
            "type": "integer",
            "format": "int64"
          },
          "analysisDate": {
            "type": "string",
            "format": "date"
          },
          "stationId": {
            "type": "integer",
            "format": "int64"
          },
          "evaluationDate": {
            "type": "string",
            "format": "date"
          },
          "totRes": {
            "type": "integer",
            "format": "int64"
          },
          "resOk": {
            "type": "integer",
            "format": "int64"
          },
          "resKoReal": {
            "type": "integer",
            "format": "int64"
          },
          "resKoValid": {
            "type": "integer",
            "format": "int64"
          },
          "kpiB9DetailResultId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}